+++ changelog.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 2023-01-25
+
+### changed
+- fix some bugs
+
+### Added
+- add test for machine and machine instance views
+
+
 ## 2023-01-21
 
 ### Added